qcmdpc_markov_ncw

Module code

Structures needed for working with QC-MDPC codes

This module provides:

Core MDPC code parameters including:

Block length (r)

Code length (n)

Column weight (d)

Row weight (w)

Degree distributions of nodes in the subgraph G of the Tanner graph given by near-codewords of the form xⁱh0(x)⊕0 and 0⊕xⁱh1(x) (see 2.3 of [ALMPRTV25])

Reference

[ALMPRTV25]: Sarah Arpin, Jun Bo Lau, Antoine Mesnard, Ray Perlner, Angela Robinson, Jean-Pierre Tillich & Valentin Vasseur: Error floor prediction with Markov models for QC-MDPC codes. https://eprint.iacr.org/2025/153

Structs

CodeDegrees

Represents the degree distribution of parity check nodes in the subgraph G of the Tanner graph induced by near-codewords in a QC-MDPC code.

MDPCCode

Structure for an MDPC code.

Type Aliases

BlockDegrees

Degree distribution of a single block