qcmdpc_markov_ncw::distribution

Trait ConvolveAll

pub trait ConvolveAll<T>: Iterator<Item = T>
where
    T: Convolve + Clone + Sync + Send,
{
    // Required method
    fn convolve_all(self, max_idx: usize) -> Option<T>;
}

Convolve all distributions in an iterator.

Required Methods

fn convolve_all(self, max_idx: usize) -> Option<T>

Convolves all elements in this iterator together up to max_idx.

Arguments

max_idx - Upper bound for indices

Returns

The resulting convolved distribution, or None if iterator is empty