qcmdpc_markov_ncw::distribution

Trait ScalarProduct

pub trait ScalarProduct<S> {
    // Required methods
    fn scalar_product(&self, other: &Self) -> Option<f64>;
    fn max_product<const K: usize>(&self, other: &Self) -> Vec<(f64, S)>;
}

Computes the scalar product between distributions.

Required Methods

fn scalar_product(&self, other: &Self) -> Option<f64>

Calculates the scalar product of two distributions.

Arguments

other - Distribution to compute product with

Returns

Some(product) or None if computation fails

fn max_product<const K: usize>(&self, other: &Self) -> Vec<(f64, S)>

Calculates the top K products between distributions and returns associated states.

Arguments

other - Distribution to compute top products with

K - Number of top products to compute

Returns

Result containing Vec of (product, state) tuples for top K products or error message if computation fails

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.