qcmdpc_markov_ncw

Module models

Core of the computation in the Markov chain.

We follow different approaches:

STModel: The simple model using only syndrome weight (s) and error weight (t) as state

STUBSpecificModel: The refined model for specific keys, using (s, t, u) where u represents the number of common bits between the error pattern and the near codeword

Modules

basic

This module provides the basis models that other derived transition models can build upon.

traits

Traits for the Markov chain models.

Structs

STInitialStateModel

Initial state model for ST (syndrome weight, error weight) states.

STModel

Simple model using syndrome weight (s) and error weight (t) states.

STUBModel

STUB model using syndrome weight (s), error weight (t), shared bits (u), and block (b) states.

STUBSpecificCounterModel

Counter model for STUB states

STUBSpecificInitialStateModel

Initial states distribution model for STUB states

Enums

STState

State of an MDPC code decoding process using syndrome weight (s) and error weight (t).

STUBState

State of an MDPC code decoding process using syndrome weight (s), error weight (t), and the number of common bits with the nearest near codeword (u), along with a block index (b).