qcmdpc_markov_ncw::models

Module basic

This module provides the basis models that other derived transition models can build upon.

The available base models track:

Syndrome weight (s) and error weight (t): A simple model

STU Specific: Building on (s, t, u) for keys with known Tanner graph near codeword degree distributions.

Structs

CountersST

Counters for good and bad positions in the MDPC code

CountersSTU

Counters for good, suspicious, erroneous, and bad positions in the MDPC code

STBasic

Basic state representation for MDPC decoding with syndrome weight (s) and error weight (t)

STBasicCounterModel

Basic counter model for ST (syndrome weight, error weight) states.

STBasicInitialStateModel

Basic initial state model for ST (syndrome weight, error weight) states.

STUBasic

Basic state representation for MDPC decoding with syndrome weight (s), error weight (t), and common bits with nearest near codeword (u)

STUBasicSpecificCounterModel

Basic counter model for STU states with specific degree distributions.

STUBasicSpecificInitialStateModel

Basic initial state model for STU states with specific degree distributions.