qcmdpc_markov_ncw::models

Module traits

Traits for the Markov chain models.

Traits

Counter

Counters associated with states in MDPC code decoding.

CounterModel

Defines an interface for counter-based models in MDPC decoding. Provides methods for calculating counter probability distributions based on decoder states.

InitialStateModel

Defines an interface for initial states distribution in MDPC decoding. Provides methods for calculating state probability distributions based on error weights.

State

State in the MDPC decoding process.

TransitionModel

Transitions between states in the Markov chain.