qcmdpc_markov_ncw::models::traits

Trait Counter

pub trait Counter<T: Clone>: Sync {
    // Required methods
    fn lock(
        &self,
        code: &MDPCCode,
        state: T,
        threshold: usize,
    ) -> (F64Log, F64Log);
    fn uniform(&self, code: &MDPCCode, state: T) -> Result<Self>
       where Self: Sized;
    fn filter(&self, threshold: usize) -> (Self, F64Log)
       where Self: Sized;
}

Counters associated with states in MDPC code decoding.

This trait defines methods for manipulating and analyzing counters that are used in the decoding process of MDPC codes. It provides functionality for conditioning counters, computing locking probabilities, applying sampling, and computing transitions.

Type Parameters

T: A tuple type representing the state associated with these counters.

Required Methods

fn lock(&self, code: &MDPCCode, state: T, threshold: usize) -> (F64Log, F64Log)

Computes the “locking” probability for the given state and threshold.

Arguments

code - A reference to the MDPCCode.

state - The current state as a tuple.

threshold - The threshold value for decoding.

Returns

The computed locking probability as a F64Log.

fn uniform(&self, code: &MDPCCode, state: T) -> Result<Self>
where Self: Sized,

Modifies the counters based on uniform sampling.

Arguments

code - A reference to the MDPCCode.

state - The current state as a tuple.

Returns

A Result containing either the modified counters or an error message.

fn filter(&self, threshold: usize) -> (Self, F64Log)
where Self: Sized,

Computes transitions for the counters based on the given threshold.

Arguments

threshold - The threshold value for transitions.

Returns

A tuple containing the new counters, flip probability, and no-flip probability.