qcmdpc_markov_ncw::models::traits

Trait CounterModel

pub trait CounterModel {
    type State: Clone;
    type BasicState: Clone;
    type Counter: Counter<Self::BasicState>;

    // Required method
    fn get_counters_distribution(
        &self,
        state: &Self::State,
    ) -> Result<Self::Counter>;
}

Defines an interface for counter-based models in MDPC decoding. Provides methods for calculating counter probability distributions based on decoder states.

Required Associated Types

type State: Clone

The state type associated with this model.

type BasicState: Clone

type Counter: Counter<Self::BasicState>

The counter type associated with this model.

Required Methods

fn get_counters_distribution( &self, state: &Self::State, ) -> Result<Self::Counter>

Gets the probability distribution of counters for a state.

Calculates counter distributions needed for the decoding process at the current decoder state.

Arguments

state - A reference to the decoder state

Returns

Counter probability distribution for the given state, or an error if the state cannot have counters.