qcmdpc_markov_ncw::models::traits

Trait State

pub trait State:
    Clone
    + Eq
    + Hash {
    type Basic: Clone;

    // Required methods
    fn decoded_state() -> Self;
    fn to_blocked(&self) -> Self;
    fn s(&self) -> Result<usize>;
    fn t(&self) -> Result<usize>;
    fn is_absorbing(&self) -> bool;
    fn is_success(&self) -> bool;
}

State in the MDPC decoding process.

This trait defines the common interface for different types of states used in the MDPC code decoding algorithm. Implementations of this trait should provide methods for calculating counters, determining decoded and blocked states, and generating possible states from syndrome and error weights.

Every state must be able to convert to a basic representation (via to_basic()) . This basic representation provides the essential information needed for computation in the Markov chain.

Required Associated Types

type Basic: Clone

Required Methods

fn decoded_state() -> Self

Returns the decoded state.

fn to_blocked(&self) -> Self

Returns a blocked state based on the given state.

fn s(&self) -> Result<usize>

Returns the syndrome weight of the current state.

Errors

Returns an error if called on a state that does not have a syndrome weight.

fn t(&self) -> Result<usize>

Returns the error weight of the current state.

Errors

Returns an error if called on a state that does not have an error weight.

fn is_absorbing(&self) -> bool

Returns whether this state is absorbing in the Markov chain.

fn is_success(&self) -> bool

Returns whether this state is a success in the Markov chain.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.