qcmdpc_markov_ncw::models::traits

Trait TransitionModel

pub trait TransitionModel<S: State>: Send + Sync {
    // Required methods
    fn iter_all_states(&self) -> Vec<(usize, Vec<S>)>;
    fn transitions_from(
        &self,
        state: &S,
        threshold: Vec<usize>,
    ) -> Result<Vec<(S, F64Log)>>;
}

Transitions between states in the Markov chain.

Required Methods

fn iter_all_states(&self) -> Vec<(usize, Vec<S>)>

Generates an iterator of possible states for t between t_low and t_high inclusive.

fn transitions_from( &self, state: &S, threshold: Vec<usize>, ) -> Result<Vec<(S, F64Log)>>

Computes the possible next states and their transition probabilities from a given state

Arguments

state - The current state to transition from

threshold - The threshold value for transitions

Returns

A Result containing a vector of tuples with possible next states and their transition probabilities