qcmdpc_markov_ncw::serialize

Function stream_write_to_file

pub fn stream_write_to_file<I, S, T, W>(iter: I, writer: &mut W) -> Result<()>
where
    I: Iterator<Item = (S, T)>,
    S: Serialize,
    T: Serialize,
    W: Write,

Writes state-value pairs to a writer in JSON format, with each pair on a separate line.

Arguments

iter - An iterator yielding tuples of (state, value) pairs to be serialized

writer - A mutable reference to a writer where the JSON data will be written

Returns

Returns Ok(()) on success, or an io::Error if writing fails or serialization fails.

Type Parameters

I - Iterator type that yields (S, T) tuples

S - State type that implements Serialize

T - Value type that implements Serialize

W - Writer type that implements Write